+++ changelog.md
@@ -1,5 +1,14 @@
 # Vocal Raid Assistant
 
+## 3.7.0
+- Added a lot of spells
+
+## 3.6.1
+- Fix Addon not loading due to more removed API function
+
+## 3.6.0
+- Fix Addon not loading due to removed API function
+
 ## 3.5.7
 - Added Fade
 - Added Vengeful Retreat